removing reliance on pandas
