+++ ods_tools/oed/required_fields/README.md
@@ -7,12 +7,17 @@ We can check the field dependencies of a file with the following code:
     
 ```python
 from pandas import DataFrame, read_excel, read_csv
-from ods_tools.oed.required_fields.files_map import FilesMap, FileType
+import json
+from ods_tools.oed.required_fields.field_reference import FileFieldReference
 
 
 locations_data: DataFrame = read_csv("locations.csv")
-file_map = FilesMap(schema_path="../../OpenExposureData_Spec.xlsx")
-outcome = file_map.get_missing_fields(file_type=FileType("ReinsScope"), file_data=locations_data)
+with open("./oed.json") as json_file:
+    oed_schema: dict = json.load(json_file)
+    
+field_map = FileFieldReference()
+field_map.populate_from_json(json_data=oed_schema)
+outcome = field_map.get_missing_fields(file_data=locations_data)
 ```
 The outcome is a dictionary of missing fields and the fields that depend on them. For instance, let us alter the 
 data so that we can see some missing fields. We can insert a parent field that we know is not in the data under 
@@ -20,10 +25,9 @@ the name "test" with the code below:
 
 ```python
 from ods_tools.oed.required_fields.field import Field
-from ods_tools.oed.required_fields.files_map import FileType
 
 # insert a False parent that does not exist in the data loaded
-file_map.files[FileType("ReinsScope").value].code_refs["CR6"] = [
+field_map.code_refs["CR6"] = [
     Field(
         file_name="ReinsScope",
         input_field_name="test",
@@ -36,7 +40,7 @@ file_map.files[FileType("ReinsScope").value].code_refs["CR6"] = [
 ]
 
 # insert a False dependency for AccNumber field that does not exist in the data loaded
-file_map.files[FileType("ReinsScope").value].code_refs["CR6-01-1"].append(
+field_map.code_refs["CR6-01-1"].append(
     Field(
         file_name="ReinsScope",
         input_field_name="test_two",
